requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
hdrhistogram==0.10.3
iniconfig==2.0.0
kiwisolver==1.4.7
logging==0.4.9.6
matplotlib==3.9.3
mizani==0.13.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pbr==6.1.0
pillow==11.0.0
plotnine==0.14.3
pluggy==1.5.0
pyarrow==18.1.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
simpy==4.1.1
six==1.17.0
statsmodels==0.14.4
tzdata==2024.2
